Generate and provide an SSH config file to shell script being run
Currently, shell scripts being invoked have no way of finding what
machine is being provisioned, or what are the connection informations to
use. Trying to invoke ``vagrant ssh-config`` in the inline script is not
possible because Vagrant locks the machine.

This commit generates an "ssh-config" file (similar to the ssh-config
target) as a temporary file and provides the path to it in the
VAGRANT_SSH_CONFIGFILE environment variable.